Startup file refactoring
